+++ src/utils/file-reader.js
@@ -1,0 +1,9 @@
+const fs = require('fs')
+
+const Reader = () => {}
+
+Reader.read = (filepath) => {
+  return fs.readFileSync(filepath)
+}
+
+module.exports = Reader